refactor: rename packages to np.com.moco and made folder as the standard structur
