+++ src/main/java/np/com/moco/seleniumproject/Seleniumproject.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  */
 
-package com.mycompany.seleniumproject;
+package np.com.moco.seleniumproject;
 
 /**
  *
