+++ src/test/java/np/com/moco/seleniumproject/assertions/java.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package np.com.moco.seleniumproject.assertions;
+
+/**
+ *
+ * @author sandesh
+ */
+public class java {
+    
+}
