+++ src/test/java/np/com/moco/seleniumproject/base/basetest.java
@@ -1,4 +1,4 @@
-package base;
+package np.com.moco.seleniumproject.base;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
