+++ src/test/java/np/com/moco/seleniumproject/tests/user/CreateUserTest.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.CreateUserTest to edit this template
+ */
+package np.com.moco.seleniumproject.tests.user;
+
+/**
+ *
+ * @author sandesh
+ */
+public class CreateUserTest {
+    
+}
