+++ src/test/resources/testng.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="Suite">
+  <test name="Test">
+    <classes>
+      <class name="np.com.moco.seleniumproject.tests.user.CreateUserTest"/>
+    </classes>
+  </test>
+</suite>
